Consolidated all the extensions
